Fix issues with string search
We shouldn't change all string searches into regexp search. It will invalid all indexes on these fields. Tasks for converting wildcard characters like `%` to regexp are already done previously with matched modifier. Also if the schema type is `date` or `datetime`, it's better to support string type date query, they might from a blueprint query so it can't be searched if we don't convert it to `Date` object.