Fix issues with string search
We shouldn't change all string searches into regexp search. It will invalid all indexes and greatly hurts performance. Tasks for converting wildcard characters like `%` to regexp are already done previously with matched modifier so in `parseValue` method we no need to do it again. Let's say in some edge cases, when a user wants to search string with `null` or `true` or something contains '%' as a string value, he/she cannot do it since it's been converted to regexp. Also if the schema type is `date` or `datetime`, it's better to support string type date query, they might from a blueprint query so it can't be searched if we don't convert it to `Date` object.