Use MongoDB fields projection on update to reduce the memory footpints
As the discussion in another issue balderdashy/waterline#805, I would suggest for an update operation, it's not needed to return the full updated result set. But even it's for compatibility reason we keep it, it should have an option to turn it of. At lease, if nothing will change please use MongoDB projection when only fetching the needed `_id` field