Add support for wlNext flag caseSensitive
It's false by default to keep consistency with current version.
Difference is you can now turn it to true, which would prevent strings to be parsed automatically into RegExp. That was necessary because of waterline SQL origin which was case insensitive. However, turning every strings into RegExp invalid any index you can set in your mongo database, which is really bad for performance.